add spring security

+++ src/main/java/com/GroupProject/ecommerce/backend/controller/ProductController.java
@@ -2,6 +2,8 @@ package com.GroupProject.ecommerce.backend.controller;
 
 import com.GroupProject.ecommerce.backend.model.Product;
 import com.GroupProject.ecommerce.backend.service.ProductService;
+import com.GroupProject.ecommerce.backend.service.UserService;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
@@ -13,18 +15,29 @@ public class ProductController {
     private final AuthController authController;
     private final ProductService productService;
 
-    public ProductController(AuthController authController, ProductService productService) {
+    private final UserService userService;
+
+    public ProductController(AuthController authController, ProductService productService, UserService userService) {
         this.authController = authController;
         this.productService = productService;
+        this.userService = userService;
     }
+
 
     @PostMapping("/products/create")
     public Product CreateProduct(Product product){
 
-        this.productService.createProduct(product);
+        if (this.userService.getType()) {
 
-        return product;
 
+            this.productService.createProduct(product);
+
+            return product;
+
+        }else {
+
+            throw new AccessDeniedException("Access denied: You are not an admin");
+        }
 
 
     }
@@ -32,7 +45,17 @@ public class ProductController {
     @GetMapping("/products/get")
     public List<Product> CreateProduct(){
 
-       return this.productService.getProducts();
+        if (this.userService.getType()) {
+
+
+            return this.productService.getProducts();
+
+        }else {
+
+            throw new AccessDeniedException("Access denied: You are not an admin");
+        }
+
+
 
 
     }
@@ -41,19 +64,48 @@ public class ProductController {
     public Product getProduct(@PathVariable("productIndex") int productIndex)
     {
 
-        return productService.getProduct(productIndex);
+        if (this.userService.getType()) {
+
+
+            return productService.getProduct(productIndex);
+
+        }else {
+
+            throw new AccessDeniedException("Access denied: You are not an admin");
+        }
+
+
+
     }
 
     @PutMapping("products/update-product/{productIndex}")
     public Product updateProduct(@PathVariable("productIndex") int productIndex, Product product)
     {
+        if (this.userService.getType()) {
 
-        return productService.updateProduct(productIndex, product);
+
+            return productService.updateProduct(productIndex, product);
+
+        }else {
+
+            throw new AccessDeniedException("Access denied: You are not an admin");
+        }
+
+
     }
     @DeleteMapping("products/delete-product/{productIndex}")
     public void deleteProduct(@PathVariable("productIndex") int index){
 
-        productService.deleteProduct(index);
+        if (this.userService.getType()) {
+
+
+            productService.deleteProduct(index);
+
+        }else {
+
+            throw new AccessDeniedException("Access denied: You are not an admin");
+        }
+
     }
